requirements.txt
certifi
charset-normalizer
click
cmake
cycler
docker-pycreds
filelock
fonttools
fsspec
gitdb
GitPython
huggingface-hub
idna
importlib_resources
Jinja2
joblib
kiwisolver
lit
MarkupSafe
matplotlib
mpmath
networkx
nimfa
numpy
packaging
pandas
patsy
pillow
platformdirs
protobuf
psutil
pyparsing
python-dateutil
pytz
PyYAML
requests
scikit-learn
scipy
seaborn
sentry-sdk
setproctitle
six
smmap
statsmodels
sympy
threadpoolctl
timm
torch
torchaudio
torchvision
tqdm
triton
typing_extensions
tzdata
urllib3
wandb
zipp
